Migrate to Brunch
Migrated the app to Brunch to simplify the build process once Backbone
was introduced. It is the epitome of hipsterism with all the fancy
frameworks integrated, but should provide a sensible application
framework for working with Backbone.

Everything is still Coffeescript-based. See README for further details
on what Brunch is.